api/requirements.txt
azure-functions
openai
beautifulsoup4
requests
reportlab
